grc_system/requirements.txt
# Django and REST Framework
Django==4.2.27
djangorestframework==3.15.2
django-cors-headers==4.3.1
django-filter==24.3
djangorestframework-simplejwt==5.3.1

# Database
psycopg2-binary==2.9.9

# Celery for task scheduling
celery==5.4.0
redis==5.0.8
django-celery-beat==2.6.0
django-celery-results==2.5.1

# PDF Generation
weasyprint==62.3
reportlab==4.2.2

# File handling
Pillow==10.4.0
python-magic==0.4.27

# Utilities
python-dateutil==2.9.0
django-environ==0.11.2

# Development
ipython==8.26.0
